Spliting code by files
