+++ slices.go
@@ -73,40 +73,3 @@ func SliceInStringSlices(slice []string, slices [][]string) bool {
 	}
 	return false
 }
-
-// IsOdd Check if number is odd.
-func IsOdd(number int) bool {
-	return !(number%2 == 0)
-}
-
-// AverageFloat64 Calculate average value in list of float64.
-func AverageFloat64(xs []float64) float64 {
-	total := 0.0
-	for _, v := range xs {
-		total += v
-	}
-	return total / float64(len(xs))
-}
-
-// Copy map into new map.
-func CopyMap(sourceMap map[string]interface{}) map[string]interface{} {
-	newMap := make(map[string]interface{})
-	for k, v := range sourceMap {
-		vm, ok := v.(map[string]interface{})
-		if ok {
-			newMap[k] = CopyMap(vm)
-		} else {
-			newMap[k] = v
-		}
-	}
-	return newMap
-}
-
-// Copy map of string into new map.
-func CopyMapString(sourceMap map[string]string) map[string]string {
-	newMap := make(map[string]string)
-	for k, v := range sourceMap {
-		newMap[k] = v
-	}
-	return newMap
-}
