+++ slices_test.go
@@ -1,10 +1,6 @@
 package generigo
 
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-)
+import "testing"
 
 func Test_StringInSlice(t *testing.T) {
 	f := func(str string, sl []string, want bool) {
@@ -14,7 +10,7 @@ func Test_StringInSlice(t *testing.T) {
 			t.Fatalf("result mismatch; get: %v , want: %v", ok, want)
 		}
 	}
-
+	f("apple", nil, false)
 	f("apple", []string{}, false)
 	f("apple", []string{"pineapple"}, false)
 	f("apple", []string{"pineapple", "apple", "onion"}, true)
@@ -28,7 +24,7 @@ func Test_NumInSlice(t *testing.T) {
 			t.Fatalf("result mismatch; get: %v , want: %v", ok, want)
 		}
 	}
-
+	f(1, nil, false)
 	f(1, []int{}, false)
 	f(1, []int{2}, false)
 	f(1, []int{2, 3, 1}, true)
@@ -129,52 +125,4 @@ func Test_SliceInStringSlices(t *testing.T) {
 		[][]string{},
 		false,
 	)
-}
-
-func Test_IsOdd(t *testing.T) {
-	f := func(num int, want bool) {
-		t.Helper()
-		odd := IsOdd(num)
-		if odd != want {
-			t.Fatalf("result mismatch; get: %v , want: %v", odd, want)
-		}
-	}
-
-	f(4, false)
-	f(7, true)
-	f(0, false)
-}
-
-func Test_AverageFloat64(t *testing.T) {
-	f := func(floats []float64, want float64) {
-		t.Helper()
-		av := AverageFloat64(floats)
-		if av != want {
-			t.Fatalf("unexpected average; want: %v, get: %v", want, av)
-		}
-	}
-
-	f([]float64{1.1, 1.1}, 1.1)
-	f([]float64{2.0, 1.0}, 1.5)
-}
-
-func Test_CopyMap(t *testing.T) {
-	f := func(sourceMap map[string]interface{}) {
-		t.Helper()
-		newMap := CopyMap(sourceMap)
-		require.Equal(t, sourceMap, newMap)
-	}
-
-	f(map[string]interface{}{"one": 1, "two": "three"})
-}
-
-func Test_CopyMapString(t *testing.T) {
-	f := func(sourceMap map[string]string) {
-		t.Helper()
-		newMap := CopyMapString(sourceMap)
-		require.Equal(t, sourceMap, newMap)
-	}
-
-	f(map[string]string{"one": "1", "two": "three"})
-	f(map[string]string{})
 }
